timeseries_project/requirements.txt
djangorestframework~=3.14.0
pandas~=2.1.0
numpy~=1.26.0
django~=4.2.5
pyyaml~=6.0
psycopg2